+++ openldap/CHANGELOG.md
@@ -12,6 +12,7 @@
 * Improve openldap and LAM logging
 * Improve ldap clustering setup with correct server-id parameters
 * Fixing clustering setup and add entries to /etc/hosts
+* Adjust openldap to use hostnames localldap and remoteldap set in /etc/hosts
 
 ---